extract useSaveHotkey hook
